requirements.txt
matplotlib==3.9.4
Pillow==11.3.0
torch==2.8.0
torchvision==0.23
tensorboard==2.20.0
